requirements.txt
annotated-types==0.7.0
anthropic==0.63.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
distro==1.9.0
h11==0.16.0
html2text==2025.4.15
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
requests==2.32.4
sniffio==1.3.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
